Move auth connection step to connections
