+++ tests/behaviour/connection/steps.rs
@@ -19,7 +19,10 @@
  * under the License.
  */
 
+use std::path::PathBuf;
+
 use cucumber::{given, then, when};
+use typedb_client::{Connection, Credential, Options, TransactionType};
 
 use crate::{behaviour::Context, generic_step_impl};
 
@@ -30,11 +33,33 @@ generic_step_impl! {
     #[step("connection opens with default authentication")]
     async fn connection_opens_with_default_authentication(_: &mut Context) {}
 
+    #[step(expr = "connection opens with authentication: {word}, {word}")]
+    async fn connection_opens_with_authentication(context: &mut Context, login: String, password: String) {
+        context.set_connection(
+            Connection::new_encrypted(
+                &["localhost:11729", "localhost:21729", "localhost:31729"],
+                Credential::with_tls(
+                    &login.as_str(),
+                    &password.as_str(),
+                    Some(&PathBuf::from(
+                        std::env::var("ROOT_CA")
+                            .expect("ROOT_CA environment variable needs to be set for cluster tests to run"),
+                    )),
+                ).unwrap(),
+            ).unwrap()
+        );
+    }
+
     #[step("connection has been opened")]
     async fn connection_has_been_opened(_: &mut Context) {}
 
     #[step("connection does not have any database")]
     async fn connection_does_not_have_any_database(context: &mut Context) {
         assert!(context.databases.all().await.unwrap().is_empty());
+    }
+
+    #[step("connection closes")]
+    async fn connection_closes(context: &mut Context) {
+        assert!(context.connection.clone().force_close().is_ok());
     }
 }
